data/requirements_cuda.txt
setuptools
wheel

# Testing
Pillow
parameterized

# Add transformers, diffusers and scipy since it most commonly used
#diffusers
#scipy

#gradio
safetensors

--prefer-binary
--extra-index-url https://download.pytorch.org/whl/torch_stable.html
--extra-index-url https://download.pytorch.org/whl/cu117
--trusted-host https://download.pytorch.org
accelerate
albumentations
diffusers[torch]
test-tube
torch-fidelity
torch
torchvision
pytorch_lightning
transformers
picklescan
omegaconf
ftfy
regex

# GFPGAN
gfpgan
realesrgan
opencv-python

# ControlNet
controlnet_aux